Results now printed in Shiny, and loading_files split
Results are now printed to one of the output tabs.
There's still an issue with outputting the list of
questions which did not have tables automatically
generated. The message which states which questions
didn't automatically have tables generated is automatically
being repeated before each table, which is very annoying.
This commit also had some other changes as well. The
loading_files.R file had some functions which were moved
into the new reorganizing_survey_data.R file. This will be
the home for functions which reorganize the structure of the
survey and response data within the program.