Reorder qsf/csv upload, main sequence now in package
The qsf/csv upload ordering didn't make sense. It makes more sense to
upload the qsf first, then upload the responses to it.

The main sequence was taken out of the server.R function and
made into a function called get_questions_and_blocks() in loading_files.R.

The movement of the main sequence is for users who might be using the
package without the shiny app. If those users want to render
the questions and blocks with results and tables included,
there wasn't a function before this commit to do so.