I missed a transposition!
I swapped the order of the upload box -- it used to
be that the responses were uploaded first, then
the survey. That's completely illogical, so I swapped
the order so the survey was first and the responses
were second. However, I didn't update the validation
correctly. This is the commit with the updated validation
in the file upload sequence.