Separating loading files from rendering results in the interface
Before this commit, main() loaded the
survey and responses, and then processed them
into questions and blocks, which included
paired results tables and were re-organized,
pretty much ready to go for output save one
function at the end or something.

Instead, now main() simply loads the survey
and responses and returns them once they've been
loaded. The other functions and UI elements on
the page each then get the survey and responses
from the main() function and then process the
data as necessary to render the content of that
UI element.

This is better because now elements that
do not depend on all the data processing
successfully happening are not inhibited by
failures or errors in the data processing
for data they don't need.