HTML Cleaning for Results Generation
Two attempted changes:
1. stop the print(xtable...) sequence that
creates the HTML version of the results tables
from printing to console. That got wrapped
with a capture.output() to make that happen.

2. clean_html was set up so that it could
be used on any text -- now we use it!
I tried to get it working on all
question answers, question choices, and
hopefully the reports will be cleaned
of html now!