Fixing results generation
Choice data export tags have - characters
replaced with _ when they are used
for the column names in the CSV. The
link_responses_to_questions function
has been updated to reflect this.

Question choices sometimes have more elements than
only the choices' text (display logic, mostly).
This has been problematic, because I need to replace
the choice variable with the choice text, but
it turns out the first element in the choice list
is always (i hope) the display text for the choice.