Separating docx rendering.
1. Pander was removed -- that wasn't going to work.
Markdown wasn't working for storing the tables in an
intermediate transition state between R and rendering
the docx. This is because Pander was automatically
splitting tables into multiple continued tables. This
wasn't going to work for converting the document into
docx, since the tables wouldn't necessarily need to be
split the same way in docx format.

2. R/docx_rendering.R contains one function right now,
html_to_docx. This takes html, prepends and appends it
with the necessary HTML tags, and then converts it
using pandoc (on the system level, not in R), and then
returns the path to the file in the temporary folder.

3. inst/shiny/server.R was changed to use the html_to_docx
function for downloading the text_appendices and the
results_tables. This is an improvement on the previous
method, which simply saved the HTML to a file with a
'.doc' or '.xls' file extension and hoped & prayed
that the file rendered properly in MS Word/Excel.