Moved some functions, wrote get_setup and uncodeable_question_dictionary
I moved the generate_results function to the
reorganizing_survey_data.R file because the output of
the function is not HTML, and it seems that it's more
appropriate there.

I changed create_question_dictionary so that it doesn't
insert  for the SubSelector since the question
dictionary no longer includes the SubSelector column.

get_setup is a new function that sets up the
.GlobalEnv with the variables typically used
for processing survey data.

uncodeable_question_dictionary is a function
for listing a question dictionary of just
the questions that didn't have tables automatically
generated for them